+++ gateway/src/lib.rs
@@ -19,6 +19,7 @@ pub mod compact_ids;
 mod constants;
 mod env;
 mod json;
+pub mod kafka_commands;
 pub mod kafka_events;
 pub mod kafka_io;
 mod logging;
