+++ gateway/src/websocket.rs
@@ -10,6 +10,7 @@ use ws::util::Token;
 use ws::{CloseCode, Error, ErrorKind, Frame, Handler, Handshake, Message, OpCode, Result, Sender};
 
 use crate::client_events::*;
+use crate::kafka_commands::*;
 use crate::logging::*;
 use crate::model::*;
 use crate::router::RouterCommand;
